+++ src/zk/zk.prover.ts
@@ -26,7 +26,7 @@ export class ZkProver {
 
   async createWitness(input: ZkWitnessInput) {
     await this.ensureInitialized();
-    this.logger.debug(`Creating witness for userScore ${input.userScore} and minScore ${input.minScore}`);
+    this.logger.debug(`Creating witness for score ${input.score} and minScore ${input.minScore}`);
     return this.noir!.execute(input as any);
   }